Unify EStruct and ETuple
