+++ hail/src/main/scala/is/hail/expr/types/physical/PStruct.scala
@@ -47,20 +47,11 @@ final case class PStruct(fields: IndexedSeq[PField], override val required: Bool
 
   val types: Array[PType] = fields.map(_.typ).toArray
 
-  val fieldRequired: Array[Boolean] = types.map(_.required)
-
-  val fieldIdx: Map[String, Int] =
-    fields.map(f => (f.name, f.index)).toMap
-
-  val fieldNames: Array[String] = fields.map(_.name).toArray
-
   if (!fieldNames.areDistinct()) {
     val duplicates = fieldNames.duplicates()
     fatal(s"cannot create struct with duplicate ${plural(duplicates.size, "field")}: " +
       s"${fieldNames.map(prettyIdentifier).mkString(", ")}", fieldNames.duplicates())
   }
-
-  val size: Int = fields.length
 
   override def truncate(newSize: Int): PStruct =
     PStruct(fields.take(newSize), required)
@@ -76,18 +67,6 @@ final case class PStruct(fields: IndexedSeq[PField], override val required: Bool
     assert(other isOfType this)
     CodeOrdering.rowOrdering(this, other.asInstanceOf[PStruct], mb)
   }
-
-  def fieldByName(name: String): PField = fields(fieldIdx(name))
-
-  def index(str: String): Option[Int] = fieldIdx.get(str)
-
-  def selfField(name: String): Option[PField] = fieldIdx.get(name).map(i => fields(i))
-
-  def hasField(name: String): Boolean = fieldIdx.contains(name)
-
-  def field(name: String): PField = fields(fieldIdx(name))
-
-  def fieldType(name: String): PType = types(fieldIdx(name))
 
   def unsafeStructInsert(typeToInsert: PType, path: List[String]): (PStruct, UnsafeInserter) = {
     assert(typeToInsert.isInstanceOf[PStruct] || path.nonEmpty)
